Move deprecated functions to separate files.
